LayoutTests/fast/scrolling/rtl-bt-drag-scroller-expected.txt

This test makes sure that a div with right-to-left or bottom-to-top scroll directions can be scrolled by dragging the scrollbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrollMe.scrollLeft is not 0
PASS scrollMe.scrollTop is not 0
PASS successfullyParsed is true

TEST COMPLETE
